backend/jobsportal/requirements.txt
Django==5.0.1
django-filter==23.5
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-storages==1.14.2
psycopg2==2.9.9
python-dotenv==1.0.0
boto3==1.34.11
botocore==1.34.11
drf-spectacular==0.27.0
requests==2.31.0